% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-bool_expr.R
\name{bool_expr}
\alias{bool_expr}
\title{S3 Helper Function for Class 'bool_expr'}
\usage{
bool_expr(col_expr)
}
\arguments{
\item{col_expr}{character}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'bool_expr'
}
\examples{
\dontrun{
result <- bool_expr(x)
}
}
