% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-bool_formula.R
\name{bool_formula}
\alias{bool_formula}
\title{S3 Helper Function for Class 'bool_formula'}
\usage{
bool_formula(bool_formula, join_op)
}
\arguments{
\item{bool_formula}{character}

\item{join_op}{character}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'bool_formula'
}
\examples{
\dontrun{
result <- bool_formula(x)
}
}
