% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-col_formula.R
\name{col_formula}
\alias{col_formula}
\title{S3 Helper Function for Class 'col_formula'}
\usage{
col_formula(col_formula, join_op)
}
\arguments{
\item{col_formula}{character}

\item{join_op}{character}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'col_formula'
}
\examples{
\dontrun{
result <- col_formula(x)
}
}
