% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-bool_expr.R
\name{new_bool_expr}
\alias{new_bool_expr}
\title{S3 Constructor for Class 'bool_expr'}
\usage{
new_bool_expr(col_expr)
}
\arguments{
\item{col_expr}{character}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'bool_expr'
}
\examples{
\dontrun{
result <- new_bool_expr(x)
}
}
