% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-col_expr.R
\name{new_col_expr}
\alias{new_col_expr}
\title{S3 Constructor for Class 'col_expr'}
\usage{
new_col_expr(col_expr)
}
\arguments{
\item{col_expr}{character}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'col_expr'
}
\examples{
\dontrun{
result <- new_col_expr(x)
}
}
