% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-col_formula.R
\name{new_col_formula}
\alias{new_col_formula}
\title{S3 Constructor for Class 'col_formula'}
\usage{
new_col_formula(col_formula, join_op)
}
\arguments{
\item{col_formula}{character}

\item{join_op}{character}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'col_formula'
}
\examples{
\dontrun{
result <- new_col_formula(x)
}
}
