% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-bool_formula.R
\name{ns_eval.bool_formula}
\alias{ns_eval.bool_formula}
\title{S3 Method - Evaluate Expression for S3 Class 'bool_formula'}
\usage{
\method{ns_eval}{bool_formula}(obj, frame)
}
\arguments{
\item{obj}{S3 Object}

\item{frame}{data.frame}
}
\value{
character
}
\description{
S3 Method - Evaluate Expression for S3 Class 'bool_formula'
}
\examples{
\dontrun{
result <- ns_eval(bool_formula)
}
}
