% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-col_expr.R
\name{ns_eval.col_expr}
\alias{ns_eval.col_expr}
\title{S3 Method - Evaluate Expression for S3 Class 'col_expr'}
\usage{
\method{ns_eval}{col_expr}(obj, frame)
}
\arguments{
\item{obj}{S3 Object}

\item{frame}{data.frame}
}
\value{
character
}
\description{
S3 Method - Evaluate Expression for S3 Class 'col_expr'
}
\examples{
\dontrun{
result <- ns_eval(col_expr)
}
}
