% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics-col_expr.R
\name{ns_eval}
\alias{ns_eval}
\title{S3 Generic - Evaluate Expression for S3 Class 'col_expr'}
\usage{
ns_eval(obj, frame)
}
\arguments{
\item{obj}{S3 Object}

\item{frame}{data.frame}
}
\value{
R Object
}
\description{
S3 Generic - Evaluate Expression for S3 Class 'col_expr'
}
\examples{
\dontrun{
result <- ns_eval(col_expr)
result <- ns_eval(bool_col_expr)
}
}
