% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-bool_expr.R
\name{validate_bool_expr}
\alias{validate_bool_expr}
\title{S3 Validator for Class 'bool_expr'}
\usage{
validate_bool_expr(obj, bool)
}
\arguments{
\item{obj}{S3 Object}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'bool_expr'
}
\examples{
\dontrun{
test <- validate_bool_expr(obj, FALSE)
}
}
