% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-bool_formula.R
\name{validate_bool_formula}
\alias{validate_bool_formula}
\title{S3 Validator for Class 'bool_formula'}
\usage{
validate_bool_formula(obj, bool)
}
\arguments{
\item{obj}{S3 Object}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'bool_formula'
}
\examples{
\dontrun{
test <- validate_bool_formula(obj, FALSE)
}
}
