% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-col_formula.R
\name{validate_col_formula}
\alias{validate_col_formula}
\title{S3 Validator for Class 'col_formula'}
\usage{
validate_col_formula(obj, bool)
}
\arguments{
\item{obj}{S3 Object}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'col_formula'
}
\examples{
\dontrun{
test <- validate_col_formula(obj, FALSE)
}
}
